test: improve translations
