+++ src/domains/transaction/components/SecondSignatureRegistrationForm/Step4.tsx
@@ -1,0 +1,71 @@
+import { ReadWriteWallet } from "@arkecosystem/platform-sdk-profiles";
+import { BigNumber } from "@arkecosystem/platform-sdk-support";
+import { Address } from "app/components/Address";
+import { Avatar } from "app/components/Avatar";
+import { Circle } from "app/components/Circle";
+import { Icon } from "app/components/Icon";
+import { Label } from "app/components/Label";
+import { TransactionDetail } from "app/components/TransactionDetail";
+import { NetworkIcon } from "domains/network/components/NetworkIcon";
+import { TotalAmountBox } from "domains/transaction/components/TotalAmountBox";
+import React, { useEffect } from "react";
+import { useFormContext } from "react-hook-form";
+import { useTranslation } from "react-i18next";
+
+export const ReviewStep = ({ wallet }: { wallet: ReadWriteWallet }) => {
+	const { t } = useTranslation();
+	const { getValues, unregister } = useFormContext();
+	const { fee } = getValues();
+
+	useEffect(() => {
+		unregister("mnemonic");
+	}, [unregister]);
+
+	return (
+		<section data-testid="SecondSignature__review-step">
+			<h1 className="mb-0">{t("TRANSACTION.PAGE_SECOND_SIGNATURE.REVIEW_STEP.TITLE")}</h1>
+			<div className="text-theme-neutral-dark">
+				{t("TRANSACTION.PAGE_SECOND_SIGNATURE.REVIEW_STEP.DESCRIPTION")}
+			</div>
+
+			<div className="mt-4 grid grid-flow-row gap-2">
+				<TransactionDetail
+					border={false}
+					label={t("TRANSACTION.NETWORK")}
+					extra={<NetworkIcon coin={wallet.coinId()} network={wallet.networkId()} />}
+				>
+					<div className="flex-auto font-semibold truncate text-md text-theme-neutral-800 max-w-24">
+						{wallet.network().name()}
+					</div>
+				</TransactionDetail>
+
+				<TransactionDetail extra={<Avatar size="lg" address={wallet.address()} />}>
+					<div className="mb-2 font-semibold text-theme-neutral">
+						<span className="mr-1 text-sm">{t("TRANSACTION.SENDER")}</span>
+						<Label color="warning">
+							<span className="text-sm">{t("TRANSACTION.YOUR_ADDRESS")}</span>
+						</Label>
+					</div>
+					<Address address={wallet.address()} walletName={wallet.alias()} />
+				</TransactionDetail>
+
+				<TransactionDetail
+					label={t("TRANSACTION.TYPE")}
+					extra={
+						<div>
+							<Circle className="border-black bg-theme-background" size="lg">
+								<Icon name="Key" width={20} height={20} />
+							</Circle>
+						</div>
+					}
+				>
+					{t("TRANSACTION.PAGE_SECOND_SIGNATURE.REVIEW_STEP.TYPE")}
+				</TransactionDetail>
+
+				<div className="mt-2">
+					<TotalAmountBox amount={BigNumber.ZERO} fee={BigNumber.make(fee)} />
+				</div>
+			</div>
+		</section>
+	);
+};
